# fix: some wallet QOL fixes