--- a/.gitattributes
+++ b/.gitattributes
@@ -1,2 +1,6 @@
 # Set the default behavior, in case people don't have core.autocrlf set.
 * text=auto
+
+# Don't render text diffs for yarn plugins and releases
+/.yarn/plugins/** binary
+/.yarn/releases/** binary