--- a/apps/extension/src/core/domains/app/handler.ts
+++ b/apps/extension/src/core/domains/app/handler.ts
@@ -111,14 +111,13 @@ export default class AppHandler extends ExtensionHandler {
   }
 
   private async authenticate({ pass }: RequestLogin): Promise<boolean> {
-    await new Promise((resolve) =>
-      setTimeout(resolve, process.env.NODE_ENV === "production" ? 1000 : 0)
-    )
+    if (!(DEBUG || TEST)) await sleep(1000)
 
     try {
-      const transformedPassword = await this.stores.password.transformPassword(pass)
       const { secret, check } = await this.stores.password.get()
       if (!secret || !check) {
+        const transformedPassword = await this.stores.password.transformPassword(pass)
+
         // attempt to log in via the legacy method
         const primaryAccount = getPrimaryAccount(true)
         assert(primaryAccount, "No primary account, unable to authorise")
@@ -132,7 +131,7 @@ export default class AppHandler extends ExtensionHandler {
         pair.lock()
 
         // we can now set up the auth secret
-        await this.stores.password.setPlaintextPassword(pass)
+        await this.stores.password.setPassword(transformedPassword)
         await this.stores.password.setupAuthSecret(transformedPassword)
       } else {
         await this.stores.password.authenticate(pass)