--- a/apps/extension/src/core/page.ts
+++ b/apps/extension/src/core/page.ts
@@ -42,7 +42,7 @@ const enable = async (origin: string): Promise<Injected> => {
 const isTalismanHostname = (hostname: string) =>
   hostname === TALISMAN_WEB_APP_DOMAIN ||
   (DEBUG && hostname.endsWith(".talisman.pages.dev")) ||
-  (DEBUG && hostname === "localhost")
+  (DEBUG && ["localhost", "127.0.0.1"].includes(hostname))
 
 function inject() {
   // inject substrate wallet provider