Some commits which have been sitting around in my local dev env:

- https://github.com/TalismanSociety/talisman/commit/505e61b20eba2c6166494c4d20d77aa3c9a5cb35 adds 127.0.0.1 to the list of hosts we should inject `talismanSub` into for webapp development
- https://github.com/TalismanSociety/talisman/commit/ffd67e775502480d0568a208dafbb2e02bd87516 goes through and removes all hardcoded webapp URLs, moving them into the constants.ts file (which will soon be mostly offloaded to [talisman-config](https://github.com/talismansociety/talisman-config))
- https://github.com/TalismanSociety/talisman/commit/b3d8c8bb073407619b994d4769349871a0bdd06e stops github from rendering the full diff of the yarn sourcecode when we upgrade yarn / any yarn plugins
- https://github.com/TalismanSociety/talisman/commit/28d059336dc6b07baedb69460d3b528c54ce0cca fixes some performance bugs I noticed in the login flow, details below:

**In the backend**
We were hashing the password twice for every login by:
1. Calling `stores.password.transformPassword` - which we only use the result of when migrating users to the new login structure
2. Calling `stores.password.authenticate`, which internally calls `stores.password.transformPassword` as well

I ran a performance test before changing anything, and for me the double-call to the hash function caused a 2x increase in waiting before the authorisation completed (which makes sense). From approx 1000ms to 2000ms on my machine.

**In the frontend**
Only applies to the dev autologin, but annoyingly it would call the authorise backend function three times on load!
This would cause the backend to run the same hash three times in parallel - slowing the completion of all three of the authorisation calls.
On my machine this makes a 3x difference to the time spent waiting - from approx 1000ms for a single hash call, to approx 3000ms.

With the two fixes combined, this PR brings the login time from about 6000ms to 1000ms.